Estimate the ADF/SADF/GSADF test statistic

Calculates recursive and non-recursive right-tailed unit root tests statistics.

rtadf(y, r0, test = c("adf", "sadf", "gsadf"), type = c("none", "drift",
  "trend"), lags = 1, selectlags = c("Fixed", "AIC", "BIC"))

Arguments

y
Vector to be tested for an explosive root.

r0
Minimal Window size.

test
Test type, either "adf", "sadf" of "gsadf".

type
Deterministic terms, either, either "none", "drift" or "trend".

lags
Number of lags for y to be included.

selectlags
Lag selection can be achieved according to the Akaike "AIC" or the Bayes "BIC" information criteria. The maximum number of lags considered is set by lags. The default is to use a "fixed" lag length set by lags.

Value

List with test statistic, date-stamping sequence

References

Phillips, P. C. B., Wu, Y., & Yu, J. (2011). Explosive Behavior in the 1990s Nasdaq: When Did Exuberance Escalate Asset Values?, International Economic Review, 201(1), 201--226.

Phillips, P. C. B., Shi, S., & Yu, J. (2015). Testing for multiple bubbles: Historical episodes of exuberance and collapse in the S&P 500. International Economic Review, 56(4), 1034--1078.

Examples

t  <- 100
y  <- rnorm(t)
r0 <- round(t*(0.01+1.8/sqrt(t))) # minimal window size
rtadf(y, r0, test ="sadf")

#> $testStat
#> [1] -2.845611
#> 
#> $testSeq
#>   [1]        NA        NA        NA        NA        NA        NA        NA
#>   [8]        NA        NA        NA        NA        NA        NA        NA
#>  [15]        NA        NA        NA        NA -2.845611 -2.962393 -3.069983
#>  [22] -3.178152 -3.334821 -3.507838 -3.488751 -3.735607 -4.086355 -4.188658
#>  [29] -4.144043 -3.905135 -4.336086 -4.438325 -4.506458 -4.328225 -4.437199
#>  [36] -4.612012 -4.459419 -4.533838 -4.630131 -4.712298 -4.780817 -4.934462
#>  [43] -5.001980 -5.052667 -5.122090 -4.712582 -4.966174 -5.013808 -5.070491
#>  [50] -5.090060 -5.325440 -5.382003 -5.456267 -5.516554 -5.572698 -5.608973
#>  [57] -5.556386 -5.801160 -5.854234 -6.005693 -6.017954 -6.157096 -6.132478
#>  [64] -6.191096 -6.288127 -6.176227 -6.471504 -6.516493 -6.532374 -6.564288
#>  [71] -6.339686 -6.624039 -6.758409 -6.801009 -6.644011 -6.637201 -6.634453
#>  [78] -6.709752 -6.752331 -6.797451 -6.683811 -6.499152 -6.510040 -6.527590
#>  [85] -6.640657 -6.640392 -6.572922 -6.598442 -6.867047 -6.911572 -6.945343
#>  [92] -6.938110 -6.853645 -6.968786 -6.985652 -6.874317 -7.028427 -6.991871
#>  [99] -6.913288 -7.072469
#> 
#> $Cval
#>   significance.level critical.values
#> 1                90%       0.9828153
#> 2                95%       1.2957908
#> 3                99%       1.8902437
#>