Simulate critical values and date-stamping threshold sequence in parallel

Simulates in parallel critical values and date-stamping sequences for the right-tailed ADF, SADF and GSADF tests.

rtadfSimPar(t, r0, nrep = 1000, test = c("adf", "sadf", "gsadf"))

Arguments

t
Number of observations (i.e., length of the sample)

r0
Minimal window size

nrep
Number of replications

test
Test type, either "adf", "sadf" of "gsadf".

Value

List with 90, 95, 99 precent critical values and a threshold series for the date-stamping procedure

References

Phillips, P. C. B., Wu, Y., & Yu, J. (2011). Explosive Behavior in the 1990s Nasdaq: When Did Exuberance Escalate Asset Values?, International Economic Review, 201(1), 201--226.

Phillips, P. C. B., Shi, S., & Yu, J. (2015). Testing for multiple bubbles: Historical episodes of exuberance and collapse in the S&P 500. International Economic Review, 56(4), 1034--1078.

Examples

cv <- rtadfSimPar(t = 100, r0 = 10, nrep = 100, test = "sadf")

#> Error in structure(list(args = (1:nrep)(.doRNG.stream = list(c(407L, 1527642031L, -1754121692L, 1301308117L, -730270318L, 1055330283L, 717336112L), c(407L, -2091608342L, 1581389537L, 633518317L, -1834119408L, -1886801988L, -1820203918L), c(407L, -787718707L, 412677994L, 1029260306L, -384695145L, -122680089L, 1338408991L), c(407L, -795448307L, 209173862L, 1296800915L, -892137493L, -226582569L, 957664340L), c(407L, 133308396L, 670567275L, -1648322489L, -2101116606L, 1718184789L, -347524709L), c(407L, -1911302850L, 10609129L, 419981765L, -2036042574L, 1661960668L, -1575343077L), c(407L, -187597265L, 367890977L, 316885013L, 963274964L, 702864732L, 2022911206L),     c(407L, 1202998138L, -1565129781L, 848053604L, -1270349624L,     -1083700493L, -495089471L), c(407L, -1417869098L, 140576573L,     -1849589409L, -1005173311L, 61558655L, 615550697L), c(407L,     -80736929L, 1079897871L, -1698239193L, -2043887279L, 45325501L,     -1566008573L), c(407L, -517324657L, 1104345758L, 439968184L,     1899180388L, 1768850526L, 764088013L), c(407L, -1684870645L,     197723866L, -2056067423L, -882782769L, -1515365332L, 1005947646L    ), c(407L, -281781609L, -1259187147L, -2084092341L, -533662285L,     -828562556L, -1206874560L), c(407L, 1913817361L, -1834226782L,     -200642488L, 1409293785L, 1225416354L, -1940025697L), c(407L,     -422700526L, -1325576267L, 991046456L, -764483291L, -399871508L,     -1686825449L), c(407L, -756379428L, -1762431934L, -92616442L,     184163793L, -1543080496L, 557300524L), c(407L, -1495037162L,     -349687348L, 1601442413L, 1121062652L, 857603585L, 1295078766L    ), c(407L, -1310050192L, 82668972L, -48752783L, -412022485L,     994131202L, -398095887L), c(407L, 1841237081L, -744367072L,     -800275287L, 613521734L, -285294013L, -1219030909L), c(407L,     -1445063991L, -612455100L, -1173277062L, -1627111490L, -1371889842L,     -1648329990L), c(407L, 643248919L, -836669906L, 996471522L,     -789518138L, 37327875L, -1664529153L), c(407L, 1234488249L,     1915521819L, 1369826722L, -439329900L, 378232547L, 908534785L    ), c(407L, -1747587057L, 94203400L, -410589427L, 1949873881L,     -1717779960L, 1479716110L), c(407L, -547285425L, 1507357833L,     -458921303L, 29061537L, 852579212L, -1163481195L), c(407L,     -505544269L, 1021866501L, -1613970561L, 1497900214L, -532349435L,     -1643530920L), c(407L, 1891796134L, 566980651L, -899236168L,     -1623520281L, 511146291L, -1852589689L), c(407L, -2002075089L,     -1529956444L, 942991651L, -1994696987L, 2016639075L, 741936182L    ), c(407L, 939642840L, 507159790L, -350524702L, -1048984385L,     -909820886L, 979899354L), c(407L, 525172850L, 1980091251L,     1933944111L, 1712803493L, 1931510888L, -2133024146L), c(407L,     -761912543L, -749399637L, 916731629L, 83812726L, 338966638L,     -984045961L), c(407L, 584379890L, -561969447L, 1153846382L,     397291178L, 144802006L, -1054086067L), c(407L, 282492457L,     1565025430L, -1470900322L, 1472015866L, 418993399L, -796759139L    ), c(407L, 439860688L, 1650896925L, -1852307689L, 339382792L,     623971666L, -2031037880L), c(407L, -806076344L, 2095428308L,     -1515081497L, 1126060398L, 696240812L, 390920101L), c(407L,     115143012L, 784388576L, 2098429393L, 259002879L, 1902235414L,     -24908802L), c(407L, -1967709929L, 360478632L, -1253355220L,     1475189597L, 1506909639L, 794650530L), c(407L, 943990144L,     1696870075L, 1630057414L, -312803161L, 1501509938L, -756997684L    ), c(407L, 4116222L, 1470386444L, 125262832L, 355451875L,     -1531568291L, -149763200L), c(407L, 1794033294L, 913726382L,     -1324041932L, -2108791542L, 1651550395L, 865501286L), c(407L,     -1434820902L, -2132833767L, 1345787711L, -769942754L, 2145370924L,     1951122945L), c(407L, 1737602694L, 392380899L, -962260913L,     -1015152799L, -1836285128L, -1111166093L), c(407L, -2038035710L,     -885855669L, -1695440807L, -1998285508L, -1699350417L, 1465511315L    ), c(407L, -1249852704L, -974318211L, -1358965562L, -421302335L,     823475105L, 65382961L), c(407L, 220866227L, -1335232674L,     2089653065L, 1849869012L, -331879503L, 1115414212L), c(407L,     811318616L, -1214086725L, -1026219562L, 2097690449L, 264157089L,     1164344850L), c(407L, -430128369L, -9948912L, -1529228873L,     1493655748L, 319995994L, 1823230415L), c(407L, -392597062L,     2027095990L, 2026693996L, 127743801L, 1642148272L, 1865555217L    ), c(407L, -1286773673L, -1736322522L, 1541139510L, 820013893L,     18594039L, -1405062880L), c(407L, 985837269L, 898298792L,     1696763527L, -633334614L, -1043268029L, 1569001913L), c(407L,     -1130881066L, 661642715L, -1025317261L, 1293091954L, 146823648L,     2096862328L), c(407L, -547709924L, -1838862940L, -674536358L,     -191031654L, 1574729480L, -1082157020L), c(407L, -2002096257L,     1744132467L, -2099202626L, -468138397L, -867447009L, -38977284L    ), c(407L, 434617603L, -1730250505L, 1371935286L, -467426112L,     -1623810510L, -517471366L), c(407L, -887089149L, -730183643L,     2034562642L, -1978173599L, -1047375335L, 336096107L), c(407L,     -799098570L, 1919527996L, -403646767L, -1509276406L, -105450890L,     608564821L), c(407L, -573446762L, 787968027L, -1308865208L,     -772232191L, -1847260948L, -1432151477L), c(407L, 237367641L,     -1995012891L, 442814948L, 518149841L, 371487004L, -2015342644L    ), c(407L, -1410588150L, 1639199156L, -984893353L, 821159227L,     -1598168271L, 1392398926L), c(407L, 480257843L, 713777350L,     1370640610L, -1718449928L, -850231277L, -1472918192L), c(407L,     -1386071361L, -1422640466L, -616365939L, 204800938L, -2109538420L,     -1109705806L), c(407L, -1999193632L, -1601866625L, 849149576L,     948762514L, 2097723123L, 1865536294L), c(407L, -2132292632L,     2129299317L, -1379788581L, 309751975L, -1124199543L, -703123181L    ), c(407L, 952667713L, -461667980L, -414868245L, -824469927L,     200649622L, -557837020L), c(407L, -286461647L, -1195455311L,     255337209L, 1767932078L, -329409291L, -1009428542L), c(407L,     564287927L, 1751788404L, -1259638961L, -1163311347L, -1607925248L,     -621409484L), c(407L, 699289551L, -804444934L, -382227110L,     -1111515687L, -570506107L, -1981988211L), c(407L, 1247654248L,     426961122L, 453401275L, 478827450L, 1158912431L, 1019870236L    ), c(407L, -1265208095L, 1679799938L, 1421819781L, -5409339L,     1559538579L, -1226446586L), c(407L, 490422626L, 1069173895L,     256067481L, 1998999296L, -1208959336L, -2013002744L), c(407L,     200828810L, -1460119967L, -1698826254L, -1414066956L, 606333499L,     -1009874219L), c(407L, 59628463L, -1181953661L, -355139250L,     169705538L, 1612054318L, -2044232895L), c(407L, 337724603L,     310416325L, 308431123L, 993065119L, 598676456L, -779277738L    ), c(407L, -751936189L, -1867787985L, -130474814L, 1177893366L,     37028068L, 1154199484L), c(407L, -818342052L, 46398533L,     -1569350626L, -1580234465L, -1865557314L, -1745461290L),     c(407L, -1377201336L, 233541736L, 1640524732L, -1417491004L,     207827613L, 1818346511L), c(407L, -560238508L, 1188212372L,     -1283573173L, -213583851L, -1158203054L, 326388598L), c(407L,     -582956782L, -180456288L, -1038168136L, -1418488888L, 391969955L,     -829757610L), c(407L, -430519895L, -1601304095L, 350143069L,     356327895L, 797324830L, -1751097208L), c(407L, -1854255851L,     -913382843L, -1780612293L, -1832395158L, -1188559683L, -851846177L    ), c(407L, 1001714417L, 530738582L, -1613411637L, 701833752L,     -1773826223L, 56386555L), c(407L, 1788918970L, -2045880561L,     1490016131L, 1522843135L, -1061455635L, -675177027L), c(407L,     412521420L, -1867029600L, -1334137867L, 421851026L, -1924229420L,     -621210237L), c(407L, -1088360038L, 1087443297L, 893020070L,     -1073453002L, 945526135L, -1418541471L), c(407L, 629310242L,     -1378743616L, 550649125L, -1283609340L, 522027338L, 438341730L    ), c(407L, -227256247L, -1748371607L, -385060643L, -870973680L,     -14791591L, 454116104L), c(407L, 466923195L, 281057417L,     484816449L, -122217562L, 1133827843L, 1549322152L), c(407L,     1654591696L, 1727115858L, -12701709L, 1055457337L, 420184108L,     -16830714L), c(407L, 1596243008L, 54611148L, 180582098L,     -1185647021L, 996024926L, 849902930L), c(407L, 868311876L,     -909319733L, -538884462L, -2106996258L, 1653091724L, -1276261034L    ), c(407L, 1877259731L, 2017183902L, 1394098539L, 396950563L,     -1946881302L, -619624236L), c(407L, 121184397L, 265607101L,     -208704939L, -232159745L, -312622810L, -1248682440L), c(407L,     -123849519L, 1513214236L, -1265028122L, -272473336L, -661543424L,     -1406061520L), c(407L, 1646033903L, -974597224L, -1141408523L,     -471670181L, -1001668450L, -486448254L), c(407L, 1185416602L,     484775427L, -2001265734L, -1932887362L, 714878438L, 1026923521L    ), c(407L, -1846200402L, 1982932499L, 1410122449L, 990423203L,     -322152812L, -305904075L), c(407L, 68923384L, -98107817L,     -2050069268L, 1576775902L, 1016356294L, 1334354166L), c(407L,     -1829045086L, -1754568148L, 1696249411L, 1835206732L, -1426092509L,     -1966692735L), c(407L, 541963296L, 2106509485L, -1096002410L,     1505164764L, 553524378L, -148267445L), c(407L, 630909031L,     793332191L, 1394950674L, 44014533L, -1487611728L, 495891059L    ), c(407L, 742030895L, -705826472L, -1802274757L, 1697925593L,     1241460699L, 476338817L))), argnames = c("i", ".doRNG.stream"), evalenv = <environment>, specified = character(0), combineInfo = list(    fun = function (..., deparse.level = 1)     .Internal(cbind(deparse.level, ...)), in.order = FALSE, has.init = FALSE,     init = NULL, final = NULL, multi.combine = TRUE, max.combine = 100),     errorHandling = "stop", packages = c("RcppEigen", "doRNG"    ), export = NULL, noexport = NULL, options = list(), verbose = FALSE), class = "foreach") %dopar%     {        {            rngtools::RNGseed(.doRNG.stream)        }        {            teststat(t, r0, test)        }    }: could not find function "%dopar%"